+++ Changelog.md
@@ -3,9 +3,12 @@
 ### v.0.4.0 (2018-07)
 - **New:** Information for about box now all specified by assembly attributes.
 - **New:** New readable settings format for custom settings properties.
+- **Changed:** Extended `CheckForUpdatesCompleted` event.  
+	(Now also raised if check for updates failed.)
 - **Changed:** `UpdateNotifyMode` specifies showing of message box after update check.
 - **Fixed:** Different serializations for custom settings properties.
 - **Fixed:** For WPF: Little issues in about box and with display scaling.
+- **Fixed:** Bug in `PortableSettingsProvider` (see #1).
 
 #### v.0.3.1 (2018-04)
 - **New:** Support for adding custom settings properties to be managed.